+++ src/aac_metrics/utils/checks.py
@@ -2,8 +2,10 @@
 # -*- coding: utf-8 -*-
 
 import logging
+import re
 import subprocess
 
+from dataclasses import dataclass
 from functools import cache
 from pathlib import Path
 from subprocess import CalledProcessError
@@ -12,6 +14,7 @@ from typing import Any, Union
 
 pylog = logging.getLogger(__name__)
 
+VERSION_PATTERN = r"(?P<major>\d+)\.(?P<minor>\d+)\.(?P<patch>\d+).*"
 MIN_JAVA_MAJOR_VERSION = 8
 MAX_JAVA_MAJOR_VERSION = 11
 
@@ -41,42 +44,61 @@ def check_metric_inputs(
 
 
 def check_java_path(java_path: Union[str, Path]) -> bool:
-    """Returns True if the java path is valid."""
-    if not isinstance(java_path, (str, Path)):
-        return False
+    version = _get_java_version(str(java_path))
+    valid = _check_java_version(version, MIN_JAVA_MAJOR_VERSION, MAX_JAVA_MAJOR_VERSION)
+    if not valid:
+        pylog.error(
+            f"Using Java version {version} is not officially supported by aac-metrics package and will not work for METEOR and SPICE metrics."
+            f"(expected major version in range [{MIN_JAVA_MAJOR_VERSION}, {MAX_JAVA_MAJOR_VERSION}])"
+        )
+    return valid
 
-    output = ""
+
+def _get_java_version(java_path: str) -> str:
+    """Returns True if the java path is valid."""
+    if not isinstance(java_path, str):
+        raise TypeError(f"Invalid argument type {type(java_path)=}. (expected str)")
+
+    output = "INVALID"
     try:
         output = subprocess.check_output(
-            [str(java_path), "--version"],
+            [java_path, "-version"],
+            stderr=subprocess.STDOUT,
         )
         output = output.decode().strip()
-        version = output.split("\n")[0]
-        major_version = int(version.split(" ")[1].split(".")[0])
+        version = output.split(" ")[2][1:-1]
 
     except (
         CalledProcessError,
         PermissionError,
         FileNotFoundError,
     ) as err:
-        pylog.error(f"Invalid java path. (from {java_path=} and found error={err})")
-        return False
+        raise ValueError(f"Invalid java path. (from {java_path=} and found {err=})")
 
-    except (
-        IndexError,
-        ValueError,
-    ) as err:
-        pylog.error(f"Invalid java version. (found {output=} and {err=})")
-        return False
-
-    if not (MIN_JAVA_MAJOR_VERSION <= major_version <= MAX_JAVA_MAJOR_VERSION):
-        pylog.error(
-            f"Using Java version {version} is not officially supported by aac-metrics package and could not work for METEOR and SPICE metrics."
-            f"(found {major_version=} but expected in [{MIN_JAVA_MAJOR_VERSION}, {MAX_JAVA_MAJOR_VERSION}])"
+    except IndexError as err:
+        raise ValueError(
+            f"Invalid java version. (from {java_path=} and found {output=} and {err=})"
         )
-        return False
 
-    return True
+    return version
+
+
+def _check_java_version(version: str, min_major: int, max_major: int) -> bool:
+    result = re.match(VERSION_PATTERN, version)
+    if result is None:
+        raise ValueError(
+            f"Invalid Java version {version=}. (expected version with pattern={VERSION_PATTERN})"
+        )
+
+    major_version = int(result["major"])
+    minor_version = int(result["minor"])
+
+    if (
+        major_version == 1 and minor_version <= 8
+    ):  # java <= 8 use versioning "1.MAJOR.MINOR" and > 8 use "MAJOR.MINOR.PATCH"
+        major_version = minor_version
+
+    return min_major <= major_version <= max_major
 
 
 def is_mono_sents(sents: Any) -> bool: